requirements.txt
torchvision>=0.15.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
hydra-core>=1.3.0
pandas>=1.5.0
pillow>=9.0.0
torchmetrics>=0.11.0
omegaconf>=2.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
jupyter